requirements.txt
typer>=0.9.0
rich>=13.7.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pyyaml>=6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
